+++ src/app/history.service.ts
@@ -30,7 +30,9 @@ export class HistoryService {
       });
 
       // Filter out auto detected shortcuts
-      this.services = this.services.filter((service) => { return service.hostname == undefined});
+      this.services = this.services.filter((service) => {
+        return service.hostname == undefined;
+      });
     }
     return this.services;
   }
@@ -246,7 +248,7 @@ export class HistoryService {
 
   private extractName(url: string): string {
     try {
-      let name = url.toLowerCase().replace('http://', '').replace('https://', '').replace('www.','');
+      let name = url.toLowerCase().replace('http://', '').replace('https://', '').replace('www.', '');
       name = name.charAt(0).toUpperCase() + name.slice(1);
       if (name.endsWith('.com')) {
         name = name.substring(0, name.length - 4);