+++ src/app/home/home.page.ts
@@ -44,7 +44,7 @@ export class HomePage implements OnInit {
     isNative: Capacitor.isNativePlatform(),
     services: [],
     scanDisabled: false,
-    connectDisabled: false
+    connectDisabled: false,
   };
 
   constructor(
@@ -57,7 +57,7 @@ export class HomePage implements OnInit {
     private urlService: UrlService,
     private actionSheetCtrl: ActionSheetController,
     private settingsService: SettingsService
-  ) { }
+  ) {}
 
   async ngOnInit() {
     await this.urlService.setRemoteURL(undefined);
@@ -163,8 +163,6 @@ export class HomePage implements OnInit {
       }
     }
   }
-
-
 
   public async clearHistory() {
     this.vm.services = await this.historyService.clear();